requirements.txt
asgiref==3.6.0
django-cors-headers==3.14.0
django-filter==23.1
Django==3.2.18
djangorestframework==3.14.0
future==0.18.3
mysqlclient==2.1.1
pip==21.3.1
PyMySQL==1.0.3
pytz==2023.3
setuptools==60.2.0
sqlparse==0.4.3
typing-extensions==4.5.0
wheel==0.37.1
